Get rid of old files  src/detect.rs and src/info.rs
