--- src/info.rs
@@ -1,30 +1,0 @@
-use crate::lang::Lang;
-use crate::scripts::Script;
-
-const RELIABLE_CONFIDENCE_THRESHOLD: f64 = 0.9;
-
-/// Represents a full outcome of language detection.
-#[derive(PartialEq, Debug, Clone, Copy)]
-pub struct Info {
-    pub(crate) lang: Lang,
-    pub(crate) script: Script,
-    pub(crate) confidence: f64,
-}
-
-impl Info {
-    pub fn lang(&self) -> Lang {
-        self.lang
-    }
-
-    pub fn script(&self) -> Script {
-        self.script
-    }
-
-    pub fn is_reliable(&self) -> bool {
-        self.confidence >= RELIABLE_CONFIDENCE_THRESHOLD
-    }
-
-    pub fn confidence(&self) -> f64 {
-        self.confidence
-    }
-}
